+++ packages/editor/web/src/RichContentEditor/getBlockRenderMap.js
@@ -60,9 +60,11 @@ export default theme => {
 
   const blockRenderMap = map({
     'unordered-list-item': {
+      element: 'div',
       wrapper: <UnorderedListItem />,
     },
     'ordered-list-item': {
+      element: 'div',
       wrapper: <OrderedListItem />,
     },
   });