+++ CHANGELOG.md
@@ -28,6 +28,8 @@
     - [#527](https://github.com/wix-incubator/rich-content/pull/527) remove onTab (moved to keyBindingFn/handleKeyCommand)
   - `pubsub`
     - [#528](https://github.com/wix-incubator/rich-content/pull/528) rename visibleBlock to focusedBlock
+  - `common`
+    - [#534](https://github.com/wix-incubator/rich-content/pull/534) export isValidEditorData and normalization from common/src/lib
   
 </details>
 <hr/>