+++ CHANGELOG.md
@@ -22,7 +22,7 @@
   - `divider`
     - [#511](https://github.com/wix-incubator/rich-content/pull/511) divider's width set to 100% in editor&viewer when no initial state of width is given
   - `gallery-settings`
-    - [#535](https://github.com/wix-incubator/rich-content/pull/535) in the gallery settings, make the header, tabs and footer sticky
+    - [#535](https://github.com/wix-incubator/rich-content/pull/535) gallery setting's header, tabs and footer position
   - `editor`
     - [#522](https://github.com/wix-incubator/rich-content/pull/522) support legacy margin scss attributes
   ### :house: Internal